fix cart bugs

+++ database/migrations/2018_11_14_101220_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->increments('id');
             $table->text('note')->nullable();
-            $table->float('total')->nullable();
+            $table->double('total', 20, 5)->nullable();
             $table->timestamp('order_day')->nullable();
             $table->integer('payment_status')->nullable();
             $table->timestamp('payment_day')->nullable();
